requirements.txt
prometheus_client
python-dotenv
web3
docker
websockets
setuptools<81